Move scalabilitytests to e2e
