+++ .gitignore
@@ -26,7 +26,7 @@ package-lock.json
 .launches
 .vscode
 test-output/*
-src/client/java/teammates/client/scripts/scalabilitytests/data/
+src/e2e/java/teammates/e2e/cases/scalability/data/
 *.exe
 
 coverage/*
