+++ src/client/java/teammates/client/util/LoopHelper.java
@@ -1,4 +1,4 @@
-package teammates.client.scripts.util;
+package teammates.client.util;
 
 /**
  * Helper class to help manage loops.
