+++ src/client/java/teammates/client/util/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains utility classes for scripts dealing with the back end data.
  */
-package teammates.client.scripts.util;
+package teammates.client.util;
